Create copies of wabt's test running scripts
For now, we will use the same testing framework that WABT uses to avoid
having to write our own from scratch. However, we will very likely want
to adapt it to better suit our needs later on. So, this commit creates a
copy of the WABT's main test running scripts for use in winter. The
copied files are:

- wabt/test/run-tests.py
- wabt/test/find_exe.py
- wabt/test/utils.py

The .gitignore file is also updated to ignore artifacts generated when
the test framework is executed.