Add a header for defining C++ strong types
In many languages, it's common practice to use a "newtype" in APIs to
denote some semantic information about a parameter or return value
through the type system and to prevent accidental confusion of values
with different units of measurement or completely different meanings.
Generally, these types provide all (or some of) the functionality of the
underlying type but require explicit conversion to and from their
underlying type to make such conversions more noticeable.

While C++ doesn't have such a facility, it is possible to create a
wrapper type such that it can be used in much the same way. This is
often called a "strong type". In order to allow this to be used in
winter, a new header has been added which provides the necessary
infrastructure.