Add a config header file
In order to enable/disable certain features, it's often useful to use a
CMake configuration header file. This is a header file generated from a
template which allows for CMake variables to be used in C++ code.

Support for this has now been added and a single flag has been added
that determines whether or not mmap should be used for memory
allocation.