Add a basic implementation of WebAssembly linear memory
A very basic set of classes has been added which can be used to model a
WebAssembly linear memory. A number of basic services are provided for
loading from and storing to linear memories safely and for attempting to
grow memories at runtime.

For efficiency reasons, mmap is used to allocate the backing memories on
systems that support the necessary functions to do so. This eliminates
the need to explicitly zero newly allocated memory and thus allows
unused portions of memory to be mapped to the system's zero page. For
systems without such support, malloc is used instead.