Rename Memory::_data to _internal
Previously, the Memory::_data field held a MemoryInternal object that
contained the actual information about the Memory. This is done to allow
core properties of Memory objects to be accessible to JITted code, which
can only access standard layout types. However, the name of this field
was somewhat confusing, since it contained a completely different type
of value than what was returned by calling Memory::data(). To better
reflect its true purpose, this field has been renamed to
Memory::_internal.

Additionally, a new Memory::internal() function has been added to return
a pointer to this field. While this structure is not part of the public
API of winter, a lot of other parts of winter will need access to a
pointer to this field to perform bookkeeping. While it would
theoretically be possible to use friend declarations to restrict access,
this would get out of hand quite quickly with the number of different
components that will need access.