Improve output from WASSERT failures
Previously, a failing WASSERT was only printing the user-provided
message, then exiting with a special exit code. This was not ideal,
since it didn't play nice with debuggers and didn't give an idea of
where the failure occurred.

The new version of WASSERT will print both the file and line number of a
failure, uses std::abort to ensure that debuggers break on an assertion
failure correctly, and allows the user to provide a printf format string
with arguments.