Implement very basic module linking infrastructure
A new set of APIs have been added to support constructing and
instantiating modules. Currently, this is done by manually populating a
winter::AbstractModule object, using that to construct a winter::Module
object, then linking that inside of a winter::ImportEnvironment into a
winter::ModuleInstance object.

This also introduces some skeleton code for handling functions at a high
level. However, the actual loading and execution of functions is not yet
implemented.