Rename mock to create_mock
Previously, both UnlinkedFunc and LinkedFunc had a static method named
mock that was used to create mock objects for the purpose of being used
during unit testing. Effectively, these objects have just enough
information to be able to link modules, but would crash if executed.

However, the name of this method was deemed to be too confusing, as a
developer making use of the winter API may not realize what this name
means. To clarify this somewhat, the method has been renamed to
create_mock.